feat: sub pipe prediction
